% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcex_edit.R
\name{mcex_edit}
\alias{mcex_edit}
\title{Reads and cleans a minicex dataset. Utilises mcex_read}
\usage{
mcex_edit(edit_path, minicex_data)
}
\arguments{
\item{edit_path}{A path pointing to your excel file with edited rows}

\item{minicex_data}{A cleaning MiniCEx table (you probably generated this with mcex_read())}
}
\value{
A dataframe containing cleaned MiniCEx data
}
\description{
Reads and cleans a minicex dataset. Utilises mcex_read
}
\examples{
dat <- mcex_edit("data/minicex edits.xlsx")

}
