% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcex_enoughtasks.R
\name{mcex_enoughtasks}
\alias{mcex_enoughtasks}
\title{Generates a dataframe of students who have not performed enough tasks}
\usage{
mcex_enoughtasks(minicex_data, week_requirement = 24)
}
\arguments{
\item{minicex_data}{A cleaned MiniCEx dataset, should be generated by mcex_read()}

\item{week_requirement}{A numeric representation of the number of tasks a student should perform. Should be generated by mcex_weekn(). Default is 24}
}
\value{
A dataframe containing the number of students who have not performed enough tasks
}
\description{
Generates a dataframe of students who have not performed enough tasks
}
\examples{
notEnoughTasks <- mcex_enoughtasks(minicex_data, week_requirement)
}
