% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcex_inglis.R
\name{mcex_inglis}
\alias{mcex_inglis}
\title{Returns the students thought to be at Inglis this week}
\usage{
mcex_inglis(file_path, sheet, what_week = 0)
}
\arguments{
\item{file_path}{A path pointing to your timetable excel file}

\item{sheet}{The sheet your timetable is contained in}

\item{what_week}{A string of format YYYY-MM-DD which denotes what week do you want to check? Defaults to Sys.Date()}
}
\value{
A dataframe containing the names, weeks and rotation of each student at each week
}
\description{
Returns the students thought to be at Inglis this week
}
\examples{
inglis <- mcex_inglis("data/timetable.xlsx")
}
