% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcex_longdatetasks.R
\name{mcex_longdatetasks}
\alias{mcex_longdatetasks}
\title{Generates a long dataframe of the dates of tasks that have been performed}
\usage{
mcex_longdatetasks(minicex_data)
}
\arguments{
\item{minicex_data}{A cleaned MiniCEx dataset, should be generated by mcex_read()}
}
\value{
A dataframe containing dates of tasks and count of task type
}
\description{
Generates a long dataframe of the dates of tasks that have been performed
}
\examples{
dateTasksLong <- mcex_longdatetasks(minicex_data)
}
