% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcex_longtasks.R
\name{mcex_longtasks}
\alias{mcex_longtasks}
\title{Creates a long dataframe of the number of tasks per student}
\usage{
mcex_longtasks(tasks)
}
\arguments{
\item{tasks}{A dataframe of tasks per student (you will likely have used mcex_tasks to generate this)}
}
\value{
A dataframe of number of tasks per student
}
\description{
Creates a long dataframe of the number of tasks per student
}
\examples{
dat <- mcex_longtasks(tasks)




}
