% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcex_matriccheck.R
\name{mcex_matriccheck}
\alias{mcex_matriccheck}
\title{Performs a check that matriculation numbers in minicex data matches the autosupplied}
\usage{
mcex_matriccheck(minicex_data)
}
\arguments{
\item{minicex_data}{A dataframe of cleaned MiniCEx data (you will likely have used mcex_read() to generate this)}
}
\value{
A table of MiniCEx entries where the matriculation numbers do not match with the form submitter
}
\description{
Performs a check that matriculation numbers in minicex data matches the autosupplied
}
\examples{
check <- mcex_matriccheck()


}
