% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcex_read.R
\name{mcex_read}
\alias{mcex_read}
\title{Reads and cleans a minicex dataset. Helper function for mcex_edit()}
\usage{
mcex_read(file_path)
}
\arguments{
\item{file_path}{A path pointing to your MiniCEx excel file}
}
\value{
A dataframe containing cleaned MiniCEx data
}
\description{
Reads and cleans a minicex dataset. Helper function for mcex_edit()
}
\examples{
dat <- mcex_read("data/minicex.xlsx")
}
