% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcex_tasks.R
\name{mcex_tasks}
\alias{mcex_tasks}
\title{Creates a dataframe of the number of tasks per student}
\usage{
mcex_tasks(minicex_data)
}
\arguments{
\item{minicex_data}{A dataframe of tidy MiniCEx data (you will likely have used mcex_read to generate this)}
}
\value{
A dataframe of number of tasks per student
}
\description{
Creates a dataframe of the number of tasks per student
}
\examples{
tasks <- mcex_tasks(minicex_data)


}
