% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcex_ttable.R
\name{mcex_ttable}
\alias{mcex_ttable}
\title{Reads and cleans a Final Year timetable file}
\usage{
mcex_ttable(file_path, sheet)
}
\arguments{
\item{file_path}{A path pointing to your timetable excel file}

\item{sheet}{The sheet your timetable is contained in}
}
\value{
A dataframe containing the names, weeks and rotation of each student at each week
}
\description{
Reads and cleans a Final Year timetable file
}
\examples{
timetable <- mcex_ttable("data/timetable.xlsx")

}
