% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcex_weekn.R
\name{mcex_weekn}
\alias{mcex_weekn}
\title{Generates the week requirement for MiniCEx analysis}
\usage{
mcex_weekn(FYStartDate = "20230605")
}
\arguments{
\item{FYStartDate}{A string of YYYYMMDD which represents the Monday FY started}
}
\value{
The number of MiniCExs that should have been completed by the current week
}
\description{
Generates the week requirement for MiniCEx analysis
}
\examples{
mcex_weekn()


}
