% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcexplot_datetasks.R
\name{mcexplot_datetasks}
\alias{mcexplot_datetasks}
\title{Generates a plot of the type of task by date of task}
\usage{
mcexplot_datetasks(longdate_tasks, date_breaks = "1 week")
}
\arguments{
\item{longdate_tasks}{A dataset of tasks by date in long form, should be generated by mcex_longtasks() of tasks a student should perform. Should be generated by mcex_weekn(). Default is 24}

\item{date_breaks}{A string which describes the scale of the x axis, default is one-week}
}
\value{
A plot describing number of tasks by date
}
\description{
Generates a plot of the type of task by date of task
}
\examples{
mcexplot_datetasks(longdate_tasks)
}
