% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcexplot_facettasks.R
\name{mcexplot_facettasks}
\alias{mcexplot_facettasks}
\title{Generates a generic plot with facets utilised in analysis where the number of tasks can be viewed via specified vars}
\usage{
mcexplot_facettasks(
  minicex_data,
  x_var = week,
  fill_var = OverallAssessorFeedback,
  facet_var = Spp
)
}
\arguments{
\item{minicex_data}{A cleaned MiniCEx dataset, should be generated by mcex_read()}

\item{x_var}{The variable for the x axis, can be either matric or week. Default is week}

\item{fill_var}{A factor from the dataset to fill the data. Usually Spp or OverallAssessorFeedback. Default is OverallAssessorFeedback}

\item{facet_var}{A factor from the dataset to use as a facet. Default is Spp}
}
\value{
A plot describing number of tasks by chosen vars with a faceting variable
}
\description{
Generates a generic plot with facets utilised in analysis where the number of tasks can be viewed via specified vars
}
\examples{
mcexplotfacet_tasks(minicex_data, x_var = matric, fill_var = OverallAssessorFeedback, facet_var = Spp)
}
