% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcex_longtasks.R
\name{mcex_longtasks}
\alias{mcex_longtasks}
\title{Creates a long dataframe of the number of tasks per student}
\usage{
mcex_longtasks(minicex_data)
}
\arguments{
\item{minicex_data}{A dataframe of cleaned MiniCEx data (you will likely have used mcex_read() to generate this)}
}
\value{
A dataframe of number of tasks per student
}
\description{
Creates a long dataframe of the number of tasks per student
}
\examples{
stuTasksLong <- mcex_longtasks(minicex_data)




}
