% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcexplot_datetasks.R
\name{mcexplot_datetasks}
\alias{mcexplot_datetasks}
\title{Generates a plot of the type of task by date of task}
\usage{
mcexplot_datetasks(minicex_data, date_breaks = "1 week")
}
\arguments{
\item{minicex_data}{A cleaned MiniCEx dataset, should be generated by minicex_read()}

\item{date_breaks}{A string which describes the scale of the x axis, default is one-week}
}
\value{
A plot describing number of tasks by date
}
\description{
Generates a plot of the type of task by date of task
}
\examples{
mcexplot_datetasks(longdate_tasks)
}
